Make BuildHierarchy.directBuildParentsOf recursion-safe
The old implementation ran into stack overflows, in case of cyclic dependencies, this new implementation changes the map key to the identity path of the projects and recurses only, if the project doesn't already have an entry in the map. This is also more efficient and given that Gradle/GradleInternal impls don't implement equals and hashCode probably also safer on the long run.